stack_dump

+++ include/stack.h
@@ -2,13 +2,14 @@
 #define STACK_H
 
 #include <stddef.h>
+#include <stdint.h>
 #include "bool.h"
 
 typedef double stack_elem_t;
 typedef struct stack_str stack_t;
 
 #ifdef STACK_DEBUG
-extern const int STACK_POISON;
+extern const uint64_t STACK_POISON;
 #endif
 
 typedef enum
@@ -32,7 +33,7 @@ struct stack_str
 
 stack_status stack_construct(stack_t *st, size_t size);
 stack_status stack_destruct(stack_t *st);
-stack_t *stack_new(size_t size);
+stack_t     *stack_new(size_t size);
 stack_status stack_delete(stack_t *st);
 
 stack_status stack_push(stack_t *st, stack_elem_t elem);
@@ -41,11 +42,11 @@ stack_status stack_back(stack_t *st, stack_elem_t *out);
 
 stack_status stack_erase(stack_t *st);
 
-bool stack_empty(stack_t *st);
-size_t stack_getsize(stack_t *st);
+bool         stack_empty(stack_t *st);
+size_t       stack_getsize(stack_t *st);
 
 stack_status stack_validate(stack_t *st);
-void stack_dump(stack_t *st);
+void         stack_dump(stack_t *st);
 
 #ifdef STACK_DEBUG
 #define STACK_VALIDATE(st) {stack_status stst = stack_validate(st); if (stst != STACK_OK) return stst;}
